feat: pss-i03-session-models-cli-fanin-006 - Prove root.BuildProcess and invocation-signature stability
