+++ pkg/transports/cli/root_session_test.go
@@ -3,7 +3,6 @@ package cli
 import (
 	"bytes"
 	"context"
-	"errors"
 	"fmt"
 	"io"
 	"os"
@@ -862,249 +861,6 @@ func TestProductionRootUsesGeneratedSessionFamilyCutover(t *testing.T) {
 			t.Fatalf("Find(%s) error = %v, want non-representative families on handwritten constructors", name, err)
 		}
 	}
-}
-
-func TestSessionCommandCompositionUsesTypedSessionsCLIAdapter(t *testing.T) {
-	t.Parallel()
-
-	called := false
-	factory := NewCommandFactory(CommandOperations{
-		ShowSession: func(cfg session.ShowConfig) error {
-			called = true
-			if cfg.SessionID != "session-beta" {
-				t.Fatalf("SessionID = %q, want session-beta", cfg.SessionID)
-			}
-			return nil
-		},
-	})
-	if factory.SessionsCLI == nil {
-		t.Fatal("SessionsCLI adapter is missing from composed factory")
-	}
-
-	root := factory.NewCommand(nil, nil, nil)
-	root.SetOut(io.Discard)
-	root.SetErr(io.Discard)
-	root.SetArgs([]string{"session", "show", "session-beta"})
-	if err := root.Execute(); err != nil {
-		t.Fatalf("execute session show: %v", err)
-	}
-	if !called {
-		t.Fatal("typed Sessions adapter was not invoked through production composition")
-	}
-}
-
-func TestSessionDataCommandsPreserveBehaviorThroughProductionComposition(t *testing.T) {
-	t.Parallel()
-
-	operationFailure := errors.New("session operation failed")
-	tests := []struct {
-		name       string
-		args       []string
-		wantError  error
-		operations func(*testing.T, error) CommandOperations
-	}{
-		{
-			name: "create",
-			args: []string{
-				"--verbose", "--json", "--server", "https://factory.example",
-				"session", "create", "--dir", "/workspace/fleet",
-				"--validate-only", "--target-kind", "named", "--target-name", "alpha",
-			},
-			wantError: operationFailure,
-			operations: func(t *testing.T, result error) CommandOperations {
-				return CommandOperations{CreateSession: func(cfg session.CreateConfig) error {
-					if cfg.Server != "https://factory.example" || cfg.Dir != "/workspace/fleet" ||
-						!cfg.ValidateOnly || cfg.TargetKind != "named" || cfg.TargetName != "alpha" ||
-						!cfg.JSON || !cfg.Verbose {
-						t.Fatalf("create config = %#v", cfg)
-					}
-					return writeSessionCompositionOutput(cfg.Output, cfg.Diagnostics, result)
-				}}
-			},
-		},
-		{
-			name:      "delete",
-			args:      []string{"--verbose", "--json", "session", "delete", "session-beta", "--port", "9444"},
-			wantError: operationFailure,
-			operations: func(t *testing.T, result error) CommandOperations {
-				return CommandOperations{DeleteSession: func(cfg session.DeleteConfig) error {
-					if cfg.SessionID != "session-beta" || cfg.Port != 9444 || !cfg.JSON || !cfg.Verbose {
-						t.Fatalf("delete config = %#v", cfg)
-					}
-					return writeSessionCompositionOutput(cfg.Output, cfg.Diagnostics, result)
-				}}
-			},
-		},
-		{
-			name: "list",
-			args: []string{
-				"--verbose", "--json", "--server", "https://factory.example",
-				"session", "list", "--scope", "live",
-			},
-			wantError: operationFailure,
-			operations: func(t *testing.T, result error) CommandOperations {
-				return CommandOperations{ListSessions: func(cfg session.ListConfig) error {
-					if cfg.Context == nil || cfg.Server != "https://factory.example" ||
-						cfg.Scope != "live" || !cfg.JSON || !cfg.Verbose {
-						t.Fatalf("list config = %#v", cfg)
-					}
-					return writeSessionCompositionOutput(cfg.Output, cfg.Diagnostics, result)
-				}}
-			},
-		},
-		{
-			name:      "show",
-			args:      []string{"--verbose", "--json", "--server", "https://factory.example", "session", "show", "session-beta"},
-			wantError: context.Canceled,
-			operations: func(t *testing.T, result error) CommandOperations {
-				return CommandOperations{ShowSession: func(cfg session.ShowConfig) error {
-					if cfg.Context == nil || cfg.Server != "https://factory.example" ||
-						cfg.SessionID != "session-beta" || !cfg.JSON || !cfg.Verbose {
-						t.Fatalf("show config = %#v", cfg)
-					}
-					return writeSessionCompositionOutput(cfg.Output, cfg.Diagnostics, result)
-				}}
-			},
-		},
-		{
-			name: "dispatches",
-			args: []string{
-				"--verbose", "--json", "--server", "https://factory.example",
-				"session", "dispatches", "dur-sess-review-001",
-				"--phase", "review", "--status", "COMPLETED",
-			},
-			wantError: operationFailure,
-			operations: func(t *testing.T, result error) CommandOperations {
-				return CommandOperations{ListSessionDispatches: func(cfg session.DispatchesConfig) error {
-					if cfg.Context == nil || cfg.Server != "https://factory.example" ||
-						cfg.SessionID != "dur-sess-review-001" || cfg.Phase != "review" ||
-						cfg.Status != "COMPLETED" || !cfg.JSON || !cfg.Verbose {
-						t.Fatalf("dispatches config = %#v", cfg)
-					}
-					return writeSessionCompositionOutput(cfg.Output, cfg.Diagnostics, result)
-				}}
-			},
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name+" success", func(t *testing.T) {
-			stdout, stderr, err := executeSessionComposition(t, test.operations(t, nil), test.args)
-			if err != nil {
-				t.Fatalf("Execute() error = %v", err)
-			}
-			if stdout != "session-ok\n" || stderr != "session-diagnostic\n" {
-				t.Fatalf("stdout = %q, stderr = %q", stdout, stderr)
-			}
-		})
-		t.Run(test.name+" failure", func(t *testing.T) {
-			stdout, stderr, err := executeSessionComposition(t, test.operations(t, test.wantError), test.args)
-			if !errors.Is(err, test.wantError) {
-				t.Fatalf("Execute() error = %v, want %v", err, test.wantError)
-			}
-			if stdout != "" || stderr != fmt.Sprintf("Error: %v\n", test.wantError) {
-				t.Fatalf("failure stdout = %q, stderr = %q", stdout, stderr)
-			}
-		})
-	}
-}
-
-func TestSessionLifecycleCommandsPreserveBehaviorThroughProductionComposition(t *testing.T) {
-	t.Parallel()
-
-	operationFailure := errors.New("session lifecycle operation failed")
-	tests := []struct {
-		name       string
-		args       []string
-		wantError  error
-		operations func(*testing.T, error) CommandOperations
-	}{
-		{
-			name: "pause default compatibility session",
-			args: []string{
-				"--verbose", "--json", "--server", "https://factory.example",
-				"session", "pause",
-			},
-			wantError: operationFailure,
-			operations: func(t *testing.T, result error) CommandOperations {
-				return CommandOperations{PauseSession: func(cfg session.LifecycleControlConfig) error {
-					if cfg.Context == nil || cfg.Server != "https://factory.example" ||
-						cfg.SessionID != "" || !cfg.JSON || !cfg.Verbose {
-						t.Fatalf("pause config = %#v", cfg)
-					}
-					return writeSessionCompositionOutput(cfg.Output, cfg.Diagnostics, result)
-				}}
-			},
-		},
-		{
-			name: "resume named session",
-			args: []string{
-				"--verbose", "--json", "--server", "https://factory.example",
-				"session", "resume", "session-beta",
-			},
-			wantError: context.Canceled,
-			operations: func(t *testing.T, result error) CommandOperations {
-				return CommandOperations{ResumeSession: func(cfg session.LifecycleControlConfig) error {
-					if cfg.Context == nil || cfg.Server != "https://factory.example" ||
-						cfg.SessionID != "session-beta" || !cfg.JSON || !cfg.Verbose {
-						t.Fatalf("resume config = %#v", cfg)
-					}
-					return writeSessionCompositionOutput(cfg.Output, cfg.Diagnostics, result)
-				}}
-			},
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name+" success", func(t *testing.T) {
-			stdout, stderr, err := executeSessionComposition(t, test.operations(t, nil), test.args)
-			if err != nil {
-				t.Fatalf("Execute() error = %v", err)
-			}
-			if stdout != "session-ok\n" || stderr != "session-diagnostic\n" {
-				t.Fatalf("stdout = %q, stderr = %q", stdout, stderr)
-			}
-		})
-		t.Run(test.name+" failure", func(t *testing.T) {
-			stdout, stderr, err := executeSessionComposition(t, test.operations(t, test.wantError), test.args)
-			if !errors.Is(err, test.wantError) {
-				t.Fatalf("Execute() error = %v, want %v", err, test.wantError)
-			}
-			if stdout != "" || stderr != fmt.Sprintf("Error: %v\n", test.wantError) {
-				t.Fatalf("failure stdout = %q, stderr = %q", stdout, stderr)
-			}
-		})
-	}
-}
-
-func executeSessionComposition(
-	t *testing.T,
-	operations CommandOperations,
-	args []string,
-) (string, string, error) {
-	t.Helper()
-	factory := NewCommandFactory(operations)
-	if factory.SessionsCLI == nil {
-		t.Fatal("SessionsCLI adapter is missing from production composition")
-	}
-	root := factory.NewCommand(nil, nil, nil)
-	var stdout, stderr bytes.Buffer
-	root.SetOut(&stdout)
-	root.SetErr(&stderr)
-	root.SetArgs(args)
-	err := root.Execute()
-	return stdout.String(), stderr.String(), err
-}
-
-func writeSessionCompositionOutput(output, diagnostics io.Writer, result error) error {
-	if result != nil {
-		return result
-	}
-	if _, err := fmt.Fprintln(output, "session-ok"); err != nil {
-		return err
-	}
-	_, err := fmt.Fprintln(diagnostics, "session-diagnostic")
-	return err
 }
 
 func TestShowSessionUsesInjectedService(t *testing.T) {
